+++ TaskManager/CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**2018-08-08**
+
+- Added ``setTaskInterval()`` method.
+- Changed most parameters to const.
+
 **2018-04-03**
 
 - Added parameter to ``startTask()`` method to change the behaviour how the